Move cache and network components (#75)
- Merge /common/cache with /datasources/cache
- Move /common/networking to /datasources/networking
