+++ src/app.module.ts
@@ -8,9 +8,9 @@ import {
 import { ChainsModule } from './chains/chains.module';
 import { BalancesModule } from './balances/balances.module';
 import { LoggerMiddleware } from './common/middleware/logger.middleware';
-import { NetworkModule } from './common/network/network.module';
+import { NetworkModule } from './datasources/network/network.module';
 import { ConfigurationModule } from './common/config/configuration.module';
-import { CacheModule } from './common/cache/cache.module';
+import { CacheModule } from './datasources/cache/cache.module';
 import { DomainModule } from './domain.module';
 
 @Module({
