+++ src/balances/balances.controller.spec.ts
@@ -6,7 +6,7 @@ import chainFactory from '../domain/chains/entities/__tests__/chain.factory';
 import {
   mockNetworkService,
   TestNetworkModule,
-} from '../common/network/__tests__/test.network.module';
+} from '../datasources/network/__tests__/test.network.module';
 import { BalancesModule } from './balances.module';
 import {
   fakeConfigurationService,
@@ -16,7 +16,7 @@ import { FiatCodesExchangeResult } from '../domain/exchange/entities/fiat-codes-
 import {
   fakeCacheService,
   TestCacheModule,
-} from '../common/cache/__tests__/test.cache.module';
+} from '../datasources/cache/__tests__/test.cache.module';
 import { DomainModule } from '../domain.module';
 import { balanceFactory } from '../domain/balances/entities/__tests__/balance.factory';
