+++ src/common/decorators/pagination.data.decorator.spec.ts
@@ -1,7 +1,7 @@
 import { PaginationDataDecorator } from './pagination.data.decorator';
 import { PaginationData } from '../pagination/pagination.data';
 import { Controller, Get, INestApplication, Module } from '@nestjs/common';
-import { fakeCacheService } from '../cache/__tests__/test.cache.module';
+import { fakeCacheService } from '../../datasources/cache/__tests__/test.cache.module';
 import { Test, TestingModule } from '@nestjs/testing';
 import * as request from 'supertest';
