+++ src/datasources/network/entities/network.request.entity.ts
