+++ src/datasources/network/entities/network.response.entity.ts
