+++ src/datasources/network/network.service.interface.ts
